materials/HCI_prototype/requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3

pillow
opencv-python
ultralytics
numpy